chore(coln-store): Move packing related logic
