+++ packages/coln-store/src/op.rs
@@ -32,5 +32,3 @@ impl Op {
         }
     }
 }
-
-
